requirements.txt
requests>=2.31.0
pydantic>=2.0.0
google-generativeai
reportlab
PyPDF2>=3.0.0




